Small refactor - start moving out static methods
