+++ events/src/main/resources/application.properties
@@ -20,6 +20,8 @@ kafka.ssl.truststore.type = PEM
 # Input queue
 mp.messaging.incoming.ingress.connector=smallrye-kafka
 mp.messaging.incoming.ingress.topic=platform.upload.announce
+# Egg queue
+#.topic=platform.inventory.events
 # runtimes
 mp.messaging.incoming.ingress.group.id=runtimes-java-general
 #mp.messaging.incoming.ingress.key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
