+++ events/src/test/java/com/redhat/runtimes/inventory/events/EventConsumerIntegrationTest.java
@@ -4,9 +4,9 @@ package com.redhat.runtimes.inventory.events;
 import static com.redhat.runtimes.inventory.events.EventConsumer.CONSUMED_TIMER_NAME;
 import static com.redhat.runtimes.inventory.events.EventConsumer.INGRESS_CHANNEL;
 import static com.redhat.runtimes.inventory.events.EventConsumer.PROCESSING_EXCEPTION_COUNTER_NAME;
-import static com.redhat.runtimes.inventory.events.EventConsumer.jvmInstanceOf;
-import static com.redhat.runtimes.inventory.events.Utils.readBytesFromResources;
-import static com.redhat.runtimes.inventory.events.Utils.readFromResources;
+import static com.redhat.runtimes.inventory.events.TestUtils.readBytesFromResources;
+import static com.redhat.runtimes.inventory.events.TestUtils.readFromResources;
+import static com.redhat.runtimes.inventory.events.Utils.jvmInstanceOf;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
